refactor(storage): rehome debate sharing models out of server handler (VAL-P4A-003)
aragora.storage.share_store reached UP into aragora.server.handlers.social.sharing
(a lazy in-function import, plus a TYPE_CHECKING import) to obtain DebateVisibility
and ShareSettings, violating the layer contract (storage must not import server).

Relocate both data models DOWN to a new stdlib-only module aragora.storage.share_models
and import them there from share_store (intra-package, no upward edge). The server
sharing handler and the aragora.server.handlers.social package re-export them from the
new home, so every existing import path keeps working with identity-equal objects (no
deprecation surface: the old module paths are unchanged and still export the symbols).

Resolves the original 'avoid circular dependency' lazy import: the new home has no
aragora imports, so share_store imports it at module scope without any cycle.

Also coerce a pre-existing untyped 'shared_by' fallback in sharing.py to str so the
changed-file mypy pre-push hook stays clean (the value could previously be None while
the SocialShare field is typed str).
