+++ aragora/server/handlers/social/sharing.py
@@ -20,11 +20,11 @@ import secrets
 import threading
 import time
 from dataclasses import dataclass, field
-from enum import Enum
 from typing import Any
 
 from aragora.rbac.decorators import require_permission
 from aragora.server.validation.schema import SHARE_UPDATE_SCHEMA, validate_against_schema
+from aragora.storage.share_models import DebateVisibility, ShareSettings
 
 from ..base import (
     BaseHandler,
@@ -40,73 +40,6 @@ logger = logging.getLogger(__name__)
 
 # Rate limiter for social share APIs (60 requests per minute)
 _share_limiter = RateLimiter(requests_per_minute=60)
-
-
-class DebateVisibility(str, Enum):
-    """Visibility level for a debate."""
-
-    PRIVATE = "private"  # Only creator can access
-    TEAM = "team"  # Organization members can access
-    PUBLIC = "public"  # Anyone with link can access
-
-
-@dataclass
-class ShareSettings:
-    """Sharing settings for a debate."""
-
-    debate_id: str
-    visibility: DebateVisibility = DebateVisibility.PRIVATE
-    share_token: str | None = None
-    created_at: float = field(default_factory=time.time)
-    expires_at: float | None = None  # None = no expiration
-    allow_comments: bool = False
-    allow_forking: bool = False
-    view_count: int = 0
-    owner_id: str | None = None
-    org_id: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        """Convert to JSON-serializable dict."""
-        return {
-            "debate_id": self.debate_id,
-            "visibility": self.visibility.value,
-            "share_token": self.share_token,
-            "share_url": self._get_share_url() if self.share_token else None,
-            "created_at": self.created_at,
-            "expires_at": self.expires_at,
-            "is_expired": self.is_expired,
-            "allow_comments": self.allow_comments,
-            "allow_forking": self.allow_forking,
-            "view_count": self.view_count,
-        }
-
-    def _get_share_url(self) -> str:
-        """Generate the share URL."""
-        # This would be configured via settings in production
-        return f"/api/v1/shared/{self.share_token}"
-
-    @property
-    def is_expired(self) -> bool:
-        """Check if the share link has expired."""
-        if self.expires_at is None:
-            return False
-        return time.time() > self.expires_at
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> ShareSettings:
-        """Create from dictionary."""
-        return cls(
-            debate_id=data["debate_id"],
-            visibility=DebateVisibility(data.get("visibility", "private")),
-            share_token=data.get("share_token"),
-            created_at=data.get("created_at", time.time()),
-            expires_at=data.get("expires_at"),
-            allow_comments=data.get("allow_comments", False),
-            allow_forking=data.get("allow_forking", False),
-            view_count=data.get("view_count", 0),
-            owner_id=data.get("owner_id"),
-            org_id=data.get("org_id"),
-        )
 
 
 MAX_SHARE_SETTINGS = 10000  # Prevent unbounded memory growth
@@ -463,7 +396,7 @@ class SharingHandler(BaseHandler):
 
         db_user = self._resolve_social_user(handler, user)
         org_id = getattr(db_user, "org_id", "") or ""
-        shared_by = getattr(db_user, "id", None) or getattr(db_user, "user_id", "")
+        shared_by = getattr(db_user, "id", None) or getattr(db_user, "user_id", "") or ""
 
         share = SocialShare(
             id=secrets.token_urlsafe(8),
